+++ lib/src/view_model.dart
@@ -266,13 +266,6 @@ abstract class VmFactory<St, T extends Widget?, Model extends Vm> {
   /// This state is final inside of the factory.
   St get state => _state;
 
-  /// Gets the store environment.
-  /// This can be used to create a global value, but scoped to the store.
-  /// For example, you could have a service locator, here, or a configuration value.
-  ///
-  /// See also: [prop] and [setProp].
-  Object? get env => _store.env;
-
   /// Gets a property from the store.
   /// This can be used to save global values, but scoped to the store.
   /// For example, you could save timers, streams or futures used by actions.
@@ -283,7 +276,7 @@ abstract class VmFactory<St, T extends Widget?, Model extends Vm> {
   /// timer.cancel();
   /// ```
   ///
-  /// See also: [setProp] and [env].
+  /// See also: [setProp].
   V prop<V>(Object? key) => _store.prop<V>(key);
 
   /// Sets a property in the store.